Refactor: EventEdit Form, Editor Component 새로 추가
