+++ src/components/eventEdit/EventEdit.ts
@@ -1,0 +1,9 @@
+import EventEditEditor from './EventEditEditor';
+import EventEditForm from './EventEditForm';
+
+const EventEdit = {
+  Form: EventEditForm,
+  Editor: EventEditEditor,
+};
+
+export default EventEdit;
