+++ src/components/eventEdit/index.ts
@@ -1,2 +1,2 @@
 // eslint-disable-next-line import/prefer-default-export
-export { default as LoginForm } from './LoginForm';
+export { default as EventEdit } from './EventEdit';
