+++ src/components/index.ts
@@ -3,3 +3,4 @@ export * from './molecules';
 export * from './organisms';
 export * from './layout';
 export * from './dashboard';
+export * from './eventEdit';
