--- src/components/organisms/editor/index.ts
@@ -1,2 +1,0 @@
-// eslint-disable-next-line import/prefer-default-export
-export { default as EventEditEditor } from './EventEditEditor';
