+++ src/components/organisms/index.ts
@@ -1,3 +1,2 @@
 /* eslint-disable import/no-cycle */
 export * from './paperContainer';
-export * from './editor';
